backend/workers-python/requirements.txt
# Kafka
kafka-python==2.0.2

# Object Storage
minio==7.2.0
boto3==1.34.0  # Alternativa S3

# Banco de dados
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

# JSON rápido
orjson==3.9.10

# Configuração
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# Utilitários
python-dateutil==2.8.2
pytz==2023.3
